add wasm support and edition 2018
